Revert "Force compare int with int, instead of relying on autoboxing"
This reverts commit 00e818bcb9a9dbf676cb3fb9ea512d1d9835a7ce.

+++ plugins/net.bioclipse.cdk.business.test/src/net/bioclipse/cdk/business/test/AbstractCDKManagerPluginTest.java
@@ -1552,10 +1552,7 @@ public abstract class AbstractCDKManagerPluginTest {
         assertEquals(1, molecule.getAtomContainer().getAtomCount());
         cdk.addExplicitHydrogens(molecule);
         assertEquals(5, molecule.getAtomContainer().getAtomCount());
-        assertEquals(
-        	0,
-        	molecule.getAtomContainer().getAtom(0).getHydrogenCount().intValue()
-        );
+        assertEquals(0, molecule.getAtomContainer().getAtom(0).getHydrogenCount());
     }
 
     @Test public void testBug691() throws Exception {
@@ -1563,10 +1560,7 @@ public abstract class AbstractCDKManagerPluginTest {
         assertEquals(12, molecule.getAtomContainer().getAtomCount());
         cdk.addExplicitHydrogens(molecule);
         assertEquals(24, molecule.getAtomContainer().getAtomCount());
-        assertEquals(
-        	0,
-        	molecule.getAtomContainer().getAtom(0).getHydrogenCount().intValue()
-        );
+        assertEquals(0, molecule.getAtomContainer().getAtom(0).getHydrogenCount());
     }
 
     @Test public void testAddImplicitHydrogens() throws Exception {
@@ -1574,10 +1568,7 @@ public abstract class AbstractCDKManagerPluginTest {
         assertEquals(1, molecule.getAtomContainer().getAtomCount());
         cdk.addImplicitHydrogens(molecule);
         assertEquals(1, molecule.getAtomContainer().getAtomCount());
-        assertEquals(
-        	4,
-        	molecule.getAtomContainer().getAtom(0).getHydrogenCount().intValue()
-        );
+        assertEquals(4, molecule.getAtomContainer().getAtom(0).getHydrogenCount());
     }
 
     @Test public void testGenerate3DCoordinates() throws Exception {
